refactor(group): 그룹 커맨드 트랜잭션을 GroupProcessor로 이관
GroupService의 모든 변경 흐름(createGroup/updateGroup/updateMember/
transferOwner/deleteGroup/leaveGroup/removeMember/create·update·deleteRole)을
GroupProcessor(@Transactional)로 이관. 매니저/오너 검증(group.isManager는
members LAZY)·board+group 2애그리거트 원자성·상세 응답 조립(interests/roles/
members 3-bag LAZY)을 모두 한 tx 안에서 수행해 LazyInit·부분 실패를 방지.

Group의 세 컬렉션은 List(bag)이라 단일 fetch join(MultipleBagFetchException)
대신 Processor tx 내 지연 순차 로딩으로 조립. GroupService는 Controller
계약(clientIp 포함 시그니처)을 유지한 채 Processor에 위임.

오케스트레이션 검증은 GroupProcessorTest로 이관, GroupServiceTest는 위임 검증.
